% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAUP.R
\docType{data}
\name{AAUP}
\alias{AAUP}
\title{1984 salaries in various professional fields}
\format{
28 rows, each of which is a professional discipline:
\itemize{
\item \code{subject} name of the discipline
\item \code{ac}: average salary (USD) for academics
\item \code{nonacsal} median salary (USD) for non-academics
\item \code{fem}: fraction of the workforce that is female
\item \code{unemp}: unemployment rate in the discipline
\item \code{nonac}: fraction of the workforce that is non-academic,
\item \code{licensed}: Does work in the profession require a license (from George Cobb's paper)
}
}
\source{
George Cobb (2011) "Teaching statistics: some important tensions" \emph{Chilean Journal of Statistics} \strong{2}(1):31-62 \href{https://soche.cl/chjs/volumes/02/01/Cobb(2011).pdf}{link}
}
\usage{
data(AAUP)
}
\description{
These data were published in the American Association of University Professsor's journal,
\emph{Academe}. There were compiled by Marcia Bellas drawing on data from the 1984 Carnegie
survey of faculty, the US National Science Foundation, the National Research Council, and the US Census
Bureau. The motivation for the work was to investigate salary "disparities among faculty whose education
and experience are comparable and whose duties are broadly similar," in particular those
due to sex. Regretably, the data do not include measures of the production of workers
in the various fields or the numbers of people employed in each field.
}
\references{
M Bellas & BF Reskin (1994) "On comparable worth" \emph{Academe} \strong{80}:83-85
}
\keyword{datasets}
