% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boston_marathon.R
\docType{data}
\name{Boston_marathon}
\alias{Boston_marathon}
\title{Winning times in the Boston Marathon}
\format{
A data frame
\itemize{
\item year
\item name: the winner's name
\item country from which the winner registered
\item time: the winning time
\item sex: female or male
\item minutes: the winning time converted to minutes
}
}
\source{
\href{https://www.baa.org/races/boston-marathon/results/champions}{Boston Athletic Association}
}
\usage{
Boston_marathon
}
\description{
The Boston marathon is the oldest continuing marathon in the US.
}
\keyword{datasets}
