% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Butterfly.R
\docType{data}
\name{Butterfly}
\alias{Butterfly}
\title{World records in the 100 & 200 m butterfly swim}
\format{
A data.frame object with one row for each world record and variables
\itemize{
\item \code{time}: the record time in seconds
\item \code{swimmer} the name of the swimmer
\item \code{date} a Date object containing the date the record was made
\item \code{place} string descripting the location
\item \code{sex}: coded as \code{F} and \code{M}
\item \code{lengths}: the total distance was divided into lengths of either 25 or 50 meters. \code{lengths} gives the number of such lengths in the total distance.
\item \code{dist}: the total distance (in meters) of the race
}
}
\source{
\href{https://en.wikipedia.org/wiki/World_record_progression_100_metres_butterfly}{Wikipedia}
}
\usage{
data(Butterfly)
}
\description{
World records in the 100 & 200 m butterfly swim
}
\keyword{datasets}
