% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calif_precip.R
\docType{data}
\name{Calif_precip}
\alias{Calif_precip}
\title{Annual precipitation in California locations}
\format{
A data.frame with 30 rows, each a weather station in California
\itemize{
\item \code{station}: the name of the station
\item \code{precip} average annual precipitation in inches
\item \code{altitude} in feet
\item \code{latitude} the station's north-south location (degrees North)
\item \code{distance}: distance in miles from the coast
\item \code{orientation}: related to the rain shadow effect of the mountains.
"W" means westward facing (toward the prevailing winds). "L" mean "leeward," that is,
facing away from the prevailing winds.
}
}
\source{
P. J. Taylor (1980)  "A Pedagogic Application of Multiple Regression Analysis: Precipitation in California" \emph{Geography} \strong{65} (3) 203-212
}
\usage{
Calif_precip
}
\description{
These data are from an article in the journal \emph{Geography} that illustrates
precipitation modeling.
}
\keyword{datasets}
