% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callback.R
\docType{data}
\name{Callback}
\alias{Callback}
\alias{Names_and_race}
\title{Resume Experiment Data}
\format{
: \code{Callback}: A data frame with 4870 rows and 4 variables. Each row is one fictitious applicant
\itemize{
\item name: first name of the fictitious job applicant
\item sex: sex of applicant (female or male)
\item callback: whether the potential employer called back to follow up. (1 = yes, 0 = no)
Another data frame, \code{Names_and_race}: which first names are associated with which race.
}

An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 36 rows and 2 columns.
}
\usage{
data("Callback")

Names_and_race
}
\description{
Data from an experimental study in which researchers sent the resumes of fictitious
job applicants to potential employers. The first names of the fictitious applicants was
set randomly to sound either Black or white.
}
\references{
\itemize{
\item Imai, Kosuke. 2017. Quantitative Social Science: An Introduction. Princeton University Press.  \href{https://press.princeton.edu/books/hardcover/9780691167039/quantitative-social-science}{URL} from whence
these data were added to \code{{LST}}. In QSS, the data are called \code{resume}.
\item Marianne Bertrand and Sendhil Mullainathan (2004) “Are Emily and Greg more employable than Lakisha and Jamal? A field experiment on labor market discrimination.” American Economic Review, vol. 94, no. 4, pp. 991–1013. doi: 10.3386/w9873
}
}
\keyword{datasets}
