% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEV.R
\docType{data}
\name{CRDS}
\alias{CRDS}
\title{Smoking and lung function among youths}
\format{
A data.frame with one row for each of 645 youngsters.
\itemize{
\item age in years
\item FEV (forced expiratory lung volume) in liters
\item height in inches
\item sex
\item smoker whether or not the youngster smokes
}
}
\source{
Cummiskey, et al. (2020) "Causal Inference in Introductory Statistics Courses"
\emph{Journal of Statistics Education} \strong{28}(1) \doi{10.1080/10691898.2020.1713936}
}
\usage{
data(CRDS)
}
\description{
Data from the Childhood Respiratory Disease Study collected
in the late 1970s to examine the
effects of smoking and exposure to second-hand smoke.
on pulmonary functions in youths.
}
\keyword{datasets}
