% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_as_function.R
\name{dag_as_function}
\alias{dag_as_function}
\title{Turn a DAG into a function}
\usage{
dag_as_function(DAG)
}
\arguments{
\item{DAG}{A DAG}
}
\value{
A function which, when evaluated with a numerical input (the random-number generator
seed), will return \code{n} rows of output from the DAG.
}
\description{
This is an experimental operation. It works just fine, but I'm not sure
whether it's needed for any purpose in teaching Math 300R.
}
