% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_draw.R
\name{dag_draw}
\alias{dag_draw}
\title{Draw a DAG}
\usage{
dag_draw(DAG, ...)
}
\arguments{
\item{DAG}{The DAG to draw}

\item{\ldots}{Additional arguments to plot.igraph()}
}
\description{
Make a simple drawing of a DAG.
}
\details{
See the igraph package for more details.
}
\examples{
dag_draw(dag03)
}
