% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_intervene.R
\name{dag_intervene}
\alias{dag_intervene}
\title{Convert a DAG to an experiment}
\usage{
dag_intervene(DAG, tilde, ..., mix = 0)
}
\arguments{
\item{DAG}{An existing DAG}

\item{tilde}{Tilde expression such as \code{y ~ binom()}. The left-hand side is the name of the variable
whose mechanism is to be changed. The right-hand side is the new mechanism for that variable.}

\item{mix}{Number between 0 and 1. Fraction of the original mechanism to mix with the experiment. This is
just for convenience, since you can specify the mixing "by hand" in \code{tilde}.}

\item{\ldots}{Additional tilde expressions to include in the DAG.}
}
\description{
By an "experiment," we mean a DAG with one of the variables set exogenously. That is, an intervention.
}
\details{
Typically, an experiment is done by taking a variable that depends on other variables in the DAG and
replacing it with an exogenous expression, that is, a formula where there are no other variables on the right-hand
side. But strictly speaking
}
\examples{
dag03
dag_intervene(dag03, x ~ binom())
dag_intervene(dag03, x ~ roll(c(-1, 0, 2)))

}
