% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_from_model}
\alias{data_from_model}
\title{Extract training data from model}
\usage{
data_from_model(model, ...)
}
\arguments{
\item{model}{the model from which to extract the training data}

\item{...}{additional arguments (not used)}
}
\description{
This typically will \emph{not} be used by an end-user.
}
\details{
not all model architectures keep track of the training data
If a model architecture isn't recognized, you'll have to add a method for that class. See vignette.
}
