% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_graph.R
\name{data_graph}
\alias{data_graph}
\title{One-step data graphics}
\usage{
data_graph(
  D,
  tilde,
  ...,
  data = NULL,
  seed = 101,
  annot = c("none", "violin", "box", "both"),
  jitter = TRUE,
  cgrad = c("yellow", "darkblue")
)
}
\arguments{
\item{D}{a data frame}

\item{tilde}{tilde expression specifying \code{y ~ x} or \code{y ~ x + color}}

\item{seed}{(optional) random seed for jittering}

\item{annot}{Statistical annotation (one of "none", "violin", "box", "both")}

\item{jitter}{If \code{TRUE}, jitter categorical and zero-one variables}

\item{cgrad}{vector of two color names to set color gradient for continuous color
variables.}
}
\description{
A simple introduction to data graphics with up to three variables.
}
\examples{
mtcars |> data_graph(mpg ~ wt + cyl)
mtcars |> mutate(cyl=as.factor(cyl)) |> data_graph(mpg ~ wt + cyl, annot="both")
}
