% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{each}
\alias{each}
\title{Evaluate an expression separately for each case}
\usage{
each(ex)
}
\arguments{
\item{ex}{an expression potentially involving other variables.}
}
\description{
Evaluate an expression separately for each case
}
