% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Econ_impact_poll.R
\docType{data}
\name{Econ_outlook_poll}
\alias{Econ_outlook_poll}
\title{SIMULATED data from an economic outlook poll}
\format{
10,000 rows with three variables: \code{age}, \code{income}, \code{pessimism}

Data from a simple SIMULATION of people's pessimism about the economic state
based on age group and income group. Nothing about the real world should be
inferred from these data; they are merely to illustrate adjusting for covariates.
}
\source{
The simulation is from "Statistical Modeling: A Fresh Approach" (2/e). Code for
it is in the file \code{system.file("SM2-simulations.R", package="LSTbook")}
}
\usage{
Econ_outlook_poll
}
\description{
SIMULATED data from an economic outlook poll
}
\keyword{datasets}
