% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_training_data.R
\name{extract_training_data}
\alias{extract_training_data}
\title{Extract the training data from a model object}
\usage{
extract_training_data(model, ...)
}
\arguments{
\item{model}{A model fitted by \code{lm()} or \code{glm()} or similar.}
}
\description{
Extract the training data from a model object
}
