% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS.R
\docType{data}
\name{FARS}
\alias{FARS}
\title{Annual summaries concerning motor-vehicle related fatalities in the US#'}
\format{
A data.frame object with one row per year from 1994 to 2016
\itemize{
\item \code{year}: The year covered by the summary
\item \code{crashes} the number of incidents in that year
\item \code{drivers} the number of drivers killed in those incidents
\item \code{passengers} the number of passengers killed in those incidents
\item \code{unknown} vehicle occupants killed whose status as driver or passenger is unknown
\item \code{motorcyclists} the number of motorcyclists killed in those incidents
\item \code{pedestrians} the number of pedestrians killed in those incidentss
\item \code{pedalcyclists} the number of non-motorized cyclist deaths
\item \code{other_nonvehicle} the number of other deaths in those incidents
\item \code{vehicle_miles} the number of miles driven by all vehicles, whether they were
involved in an incident or not. (billions of miles)
\item \code{population} the population of the US (thousands of people)
\item \code{registered_vehicles} the number of motor vehicles registered in the US (thousands)
\item \code{licenced_drivers} the number of licenced drivers in the US (thousands)
}
}
\source{
From the Fatality Analysis Reporting System
of the US Department of Transportation (DOT). The
data have been put into a tidy form from the untidy version published by the DOT,
removing columns calculated from other columns and so on.
}
\usage{
data(FARS)
}
\description{
Annual summaries concerning motor-vehicle related fatalities in the US#'
}
\keyword{datasets}
