% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lesson_worksheet.R
\name{fix_answers}
\alias{fix_answers}
\title{Replace the answer blocks with a simple "ANSWER:"}
\usage{
fix_answers(content, spacer = "@")
}
\description{
Replace the answer blocks with a simple "ANSWER:"
}
