% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germany1933vote.R
\docType{data}
\name{Germany1933vote}
\alias{Germany1933vote}
\title{Voting patterns in the 1933 German national election}
\format{
A data frame with 681 rows and 7 variables. Each row is a German precinct.
\itemize{
\item self: share of potential voters who are self-employed
\item blue: share of potential voters who are blue-collar workers
\item white: share of potential voters who are white-collar workers
\item domestic: share of potential voters who are employed domestically
\item unemployed: share of potential voters who are un-employed
\item nvoter: number of eligible voters (not clear if this include people who didn't vote)
\item nazivote: number of votes for the Nazis
}
}
\usage{
data("Germany1933vote")
}
\description{
1933 was the year that Hitler and the Nazi party came to power. The initial basis for this
was a national election in which the Nazis secured a substantial fraction of the vote. (Immediately
after the election, the Nazis burned the Reichtag (the German parliament) and started repressing their political
opposition though a campaign of imprisonment and murder.)
}
\references{
\itemize{
\item Imai, Kosuke. 2017. Quantitative Social Science: An Introduction. Princeton University Press. \href{https://press.princeton.edu/books/hardcover/9780691167039/quantitative-social-science}{URL} from whence
these data were added to \code{{LST}}. In QSS, the data are called \code{nazis}.
\item G. King, O. Rosen, M. Tanner, A.F. Wagner (2008) “Ordinary economic
voting behavior in the extraordinary election of Adolf Hitler.”
Journal of Economic History, vol. 68, pp. 951–996.#'
}
}
\keyword{datasets}
