% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hill_racing.R
\docType{data}
\name{Hill_racing}
\alias{Hill_racing}
\title{Winning times in Scottish Hill races, 2005-2017}
\format{
A data.frame object with one row for each race winner. Most races have both a
male and female winner.

"year"     "sex"      "name"     "time"     "race"     "distance" "climb"
\itemize{
\item \code{time}: the winning time in seconds
\item \code{race} the name of the race. Many races are repeated over successive years.
\item \code{year} the year the race was run
\item \code{name} the name of the winning runner
\item \code{sex}: the runner's sex, coded as \code{F} and \code{M}
\item \code{distance}: the total distance of the race in km
\item \code{climb}: the total vertical climb of the race in meters
}
}
\source{
The data were scraped from the \href{https://www.scottishhillracing.co.uk/ResultsSummary.aspx}{Scottish Hill Racing site}.
}
\usage{
data(Hill_racing)
}
\description{
Winning times in Scottish Hill races, 2005-2017
}
\keyword{datasets}
