% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_classifier.R
\name{make_classifier}
\alias{make_classifier}
\title{Construct a classifier function from a formula}
\usage{
make_classifier(tilde)
}
\arguments{
\item{tilde}{A tilde expression. The left-hand side is the name for the new column,
the right-hand side is an expression written in terms of the names in the data frame
that will be handed to the classifier function constructed by \code{make_classifier()}.}
}
\value{
A function that takes a data frame as input and returns the same data frame
but with a new column added that gives the result of the classifier calculation. The new
column will have levels "+" or "-".
}
\description{
A classifier function takes a data set as input and adds a new column showing
how each row would be classified. The classification itself is just a calculation
based on the values in that row. To create the classifier function you provide a formula
description of the calculation
}
\examples{
# If the sum of a and b is positive, the classification will be +.
cfun <- make_classifier(output ~ (a + b) > 0)
}
