% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{mix_with}
\alias{mix_with}
\title{Mix two variables together. The output will have the specified R-squared
with var1 and variance one.}
\usage{
mix_with(signal, noise = NULL, R2 = 0.5, var = 1, exact = FALSE)
}
\arguments{
\item{signal}{The part of the mixture that will be correlated with
the output.}

\item{noise}{The rest of the mixture. This will be \strong{uncorrelated}
with the output only if you specify it as pure noise.}

\item{R2}{The target R-squared.}

\item{var}{The target variance.}

\item{exact}{if \code{TRUE}, make R-squared or the target variance exactly as specified.}
}
\description{
Mix two variables together. The output will have the specified R-squared
with var1 and variance one.
}
\details{
The target R-squared and variance will be achieved only
if \code{exact=TRUE} or the sample size goes to infinity.
}
