% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eval_funs.R
\name{model_eval_fun}
\alias{model_eval_fun}
\title{Helper functions to evaluate models}
\usage{
model_eval_fun(model, data = NULL, interval = "none", level = 0.95, ...)
}
\arguments{
\item{model}{A model object of the classes permitted}

\item{data}{Usually, a data table specifying the inputs to the model. But if
not specified, the training data will be used.}

\item{interval}{One of "none", "confidence", or "prediction". Not all model
types support "prediction" or even "confidence".}

\item{level}{(default 0.95) confidence or prediction level. Must be in \verb{[0,1]}}

\item{...}{additional arguments}
}
\description{
Only used internally in \code{{LST}}
}
\details{
All of the \code{eval_} functions are ex
These functions return a numerical vector (for regression types) or
a matrix of probabilities (for classifiers)
}
