% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_type.R
\name{model_family}
\alias{model_family}
\title{Check model type against model specification and data}
\usage{
model_family(
  .data,
  .tilde,
  family = c("auto", "lm", "binomial", "poisson", "svm", "gaussian", "rlm")
)
}
\arguments{
\item{.data}{A data frame or equivalent}

\item{.tilde}{A model specification as a tilde expression}

\item{family}{Requested model type, if any.}
}
\description{
This can be used to automatically determine a model type or to determine if the specified
model type is consistent with the specification/data
}
