% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{Fits models of a variety of types}
\usage{
model_fit(
  data,
  specification,
  family = c("gaussian", "binomial", "poisson", "svm", "lm", "rlm"),
  wrangle = FALSE,
  ...
)
}
\arguments{
\item{data}{A dataframe: the training data for the model}

\item{specification}{A tilde expression specifying the model is to be fitted.}

\item{family}{The type of model to fit: \code{"gaussian"} (or \code{"lm"}), \code{"binomial"},
\code{"poisson"}, \code{"svm"}, \code{"rlm"}.}

\item{wrangle}{Evaluate on a skeleton, returning a data frame}

\item{\ldots}{other arguments}
}
\description{
At its simplest, \code{model_fit()} is a replacement for \code{group_by() |> summarized()} with
only one response variable involved.
}
\details{
You can use \code{var ~ 1} if no stratification is desired.
}
