% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plot.R
\name{model_plot}
\alias{model_plot}
\title{Plot model output for representative levels of explanatory variables}
\usage{
model_plot(
  mod,
  x,
  color = NULL,
  facet = NULL,
  facet2 = NULL,
  ...,
  data = NULL,
  nlevels = 5,
  show_data = TRUE,
  jitter = TRUE,
  interval = c("none", "prediction", "confidence"),
  level = 0.95,
  data_alpha = 0.25
)
}
\arguments{
\item{mod}{A model object as constructed by \code{lm()} or \code{glm()}}

\item{x}{Name (unquoted) of the explanatory variable to put on the x-axis}

\item{color}{Like \code{x}, but for color (if any).}

\item{facet}{Like \code{x}, but for first level of facetting (if any).}

\item{facet2}{Like \code{x}, but for second level of facetting (if any).}

\item{nlevels}{Number of levels to use for categorical variables}

\item{show_data}{Whether to show the data along with the model (default TRUE)}

\item{jitter}{Whether to jitter the plotted data (default TRUE)}

\item{interval}{One of "none", "prediction", "confidence"}

\item{level}{Level to use for interval (default: 0.95)}

\item{data_alpha}{Transparency level for data (default: 0.25)}
}
\description{
Plot model output for representative levels of explanatory variables
}
\examples{
mod <- lm(time ~ distance*sex*climb, data=Hill_racing)
model_plot(mod, x=climb, color=sex, facet=distance, nlevels=4, show_data=TRUE)
}
