% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monocacy_river.R
\docType{data}
\name{Monocacy_river}
\alias{Monocacy_river}
\title{Data on run-off from the Monocacy river at Jug Bridge, Maryland.}
\format{
A data.frame with 25 observations on the following 2 variables'
\itemize{
\item \code{precipitation}: amount of rain in inches
\item \code{runoff} runoff in inches
}
}
\source{
"Probability Concepts in Engineering" A H-S Ang and W H Tang, 2007, John Wiley based on
R.K. Linsley and J.B. Franzini (1964) \emph{Water Resources Engineering} McGraw-Hill, p.68
}
\usage{
Monocacy_river
}
\description{
When it rains, some of the water is absorbed into the ground or quickly evaporates.
Some of the water runs off (the "runoff") and is collected by streams and rivers.
These data, from a 1964 reference on water resource management,
were measured after 25 storms in the basin of the Monocacy River at
Jug Bridge, Maryland, US.
}
\details{
YOU WERER HERE, COPYING from the Rd file in man/
}
\keyword{datasets}
