% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offspring.R
\docType{data}
\name{Offspring}
\alias{Offspring}
\title{Relative sizes offspring/parent for many species}
\format{
A data.frame with 3971 rows, each for a different species
\itemize{
\item species
\item class (phylogenetic class)
\item group (phylogenetic group)
\item adult Mass of the adult (female) in grams.
\item hatchling Mass of the offspring, in grams. This applies as well to
species where the offspring is born rather than hatched.
}
}
\source{
Shai Meiri, "Endothermy, offspring size and evolution of parental provisioning in vertebrates",
\emph{Biological Journal of the Linnean Society}, \strong{128}:4, pp. 1052-6 (See Appendix S1.)
}
\usage{
data("Offspring")
}
\description{
Body mass of adult of numerous vertebrate species and newly hatched or born offspring.
}
\keyword{datasets}
