% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penguins.R
\docType{data}
\name{Penguins}
\alias{Penguins}
\title{Body measurements on penguins}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 333 rows and 8 columns.
}
\usage{
Penguins
}
\description{
This is the \code{palmerpenguins::penguins} data frame with slightly simplified
variable names.
}
\keyword{datasets}
