% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.datasim}
\alias{print.datasim}
\title{Nice printing of some internal objects}
\usage{
\method{print}{datasim}(x, ..., report_hidden = FALSE)
}
\arguments{
\item{x}{A data simulation as made by \code{datasim_make()}}

\item{report_hidden}{Show the hidden nodes (nodes whose name begins with \code{.})}

\item{\ldots}{for compatibility with generic \code{print()}}
}
\description{
Nice printing of some internal objects
}
