% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_object.R
\name{print.model_object}
\alias{print.model_object}
\title{A printing method for model objects}
\usage{
\method{print}{model_object}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Not used, but here for consistency with generic print()}
}
\description{
A printing method for model objects
}
