% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2}
\alias{R2}
\title{R-squared and allied numbers}
\usage{
R2(mod)
}
\arguments{
\item{mod}{A model generated with \code{lm()}}
}
\description{
A convenience function to calculate Rsquared and F from a model, along with
the degrees of freedom in the numerator and denominator
}
