% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_summary.R
\name{regression_summary}
\alias{regression_summary}
\alias{anova_summary}
\alias{coef_summary}
\alias{conf_interval}
\title{Generate a regression or ANOVA report in the form of a data frame.}
\usage{
regression_summary(model, ...)

anova_summary(...)

coef_summary(mod)

conf_interval(mod, level = 0.95)
}
\arguments{
\item{model}{A model as produced by \code{lm()}, \code{glm()}, and so on}
}
\description{
Generate a regression or ANOVA report in the form of a data frame.
}
\details{
These are simply wrappers around \code{broom::tidy()} used to
emphasize to students that the results are a summary in the form of a regression
report, similar to the summaries produced by \code{confint()}, \code{coef()}, etc.
}
