% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rterm.R
\name{rterm}
\alias{rterm}
\title{Generate a random term in a model.}
\usage{
rterm(df = 1, rdist = rnorm, args = list(), nrow, seed = NULL)
}
\arguments{
\item{df}{Degrees of freedom of the term: positive integer}

\item{rdist}{The random number generator to use. Default \code{rdist}.}

\item{args}{A list of parameters for the function named in \code{rdist}.}

\item{nrow}{number of rows in the model term. This is determined automatically
when used in a function like \code{lm()} or \code{model_train()}.}

\item{seed}{An (optional) seed for the random number generator: integer}
}
\description{
For demonstration purposes, create a model term with
\code{df} degrees of freedom consisting of \code{df} random vectors.
}
\examples{
lm(mpg ~ hp + rterm(df=25, seed=29323), data = mtcars) |> R2()

}
