% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shipping_losses.R
\docType{data}
\name{Shipping_losses}
\alias{Shipping_losses}
\title{Shipping losses in 1941 in the Atlantic}
\format{
12 rows, one for each month of 1941
\itemize{
\item month
\item nships number of ships sunk
\item tons gross tonnage of the ships sunk. This includes
both the ship and the cargo.
\item country whether the ships were British, or belonged to
Allied or Neutral countries.
}
}
\source{
From W.S. Churchill (1952) \emph{The Grand Alliance} a history of the Second World War. Houghton Mifflin Co. Boston. p. 782
}
\usage{
data(Shipping_losses)
}
\description{
A major theater of action in World War II was the
Atlantic ocean. Germany attempted through submarine and
aerial attacks to sink shipping supplying Britain through
the war. This table summarizes the losses in 1941.
}
\keyword{datasets}
