% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortExamples.R
\docType{data}
\name{Nats}
\alias{Nats}
\alias{Tiny}
\alias{Big}
\title{Short, simple data frames for textbook examples}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 4 columns.
}
\usage{
Nats
}
\description{
These small data frames are for illustration purposes only.
\itemize{
\item \code{Nats} made up demographic and economic data, by country and year
\item \code{Big} a simplified form of the PalmerPenguin data
\item \code{Tiny} an 8-row subset of \code{Big}
}
}
\keyword{datasets}
