% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_objects.R
\docType{data}
\name{sim_objects}
\alias{sim_objects}
\alias{sim_00}
\alias{sim_01}
\alias{sim_02}
\alias{sim_03}
\alias{sim_04}
\alias{sim_05}
\alias{sim_06}
\alias{sim_07}
\alias{sim_08}
\alias{sim_09}
\alias{sim_10}
\alias{sim_11}
\alias{sim_12}
\alias{sim_flights}
\alias{sim_medical_observations}
\alias{sim_prob_21.1}
\alias{sim_satgpa}
\alias{sim_school1}
\alias{sim_school2}
\alias{sim_vaccine}
\title{Simulations for use in \emph{Lessons in Statistical Thinking}}
\format{
An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.

An object of class \code{list} (inherits from \code{datasim}) of length 2.
}
\usage{
sim_00

sim_01

sim_02

sim_03

sim_04

sim_05

sim_06

sim_07

sim_08

sim_09

sim_10

sim_11

sim_12

sim_flights

sim_medical_observations

sim_prob_21.1

sim_satgpa

sim_school1

sim_school2

sim_vaccine
}
\description{
These \code{datasim} objects are provided.
}
\details{
They are defined in the sim_library.R file in \verb{inst/}
}
\keyword{datasets}
