% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_from_tilde.R
\name{split_tilde}
\alias{split_tilde}
\title{Evaluate a tilde expression on a data frame}
\usage{
split_tilde(tilde)
}
\arguments{
\item{tilde}{A two-sided tilde expression used for model specification}

\item{data}{A data frame}
}
\description{
Evaluate a tilde expression on a data frame
}
