% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\docType{data}
\name{STAR}
\alias{STAR}
\title{STAR Project Data}
\format{
A data frame with 6325 rows and 6 variables:
\itemize{
\item race: black or white
\item classtype: kindergarten class type: small, regular, regular with aid
\item yearssmall: number of years (0 to 4) in small classes
\item hsgrad: high-school graduation (graduated or not). NOTE: There are many NAs
\item g4math: total scaled score for the math portion of the fourth-grade standardized test
\item g4reading: total scaled score for the reading portion of the fourth-grade standardized test
}
}
\usage{
data("STAR")
}
\description{
Data from the STAR (Student–Teacher Achievement Ratio) Project, a
four-year longitudinal study examining the effect of class size in early grade
levels on educational performance and personal development
}
\references{
\itemize{
\item Imai, Kosuke. 2017. Quantitative Social Science: An Introduction. Princeton University Press. \href{https://press.princeton.edu/books/hardcover/9780691167039/quantitative-social-science}{URL} from whence
these data were added to \code{{LST}}.
\item Mosteller, Frederick. 1997. “The Tennessee Study of Class Size in the Early School Grades.” Bulletin of the American Academy of Arts and Sciences 50(7): 14-25. doi = 10.2307/3824562
}
}
\keyword{datasets}
