% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticket_sales.R
\docType{data}
\name{Ticket_sales}
\alias{Ticket_sales}
\title{Day-by-day ticket sales from OR 310}
\format{
A data frame with one row for day of the month.
\itemize{
\item \code{dow} Day of the week
\item \code{day} Day (number) in the month
\item \code{sales} Ticket sales in 1000s
}
}
\source{
Kindly provided by Operations-Research/Management professor Gerry Gonzales
}
\usage{
data(Ticket_sales)
}
\description{
OR 310 is an operations research course at the US Air Force Academy.
These data have been used in a class example for many years. Their origin is
no longer known.
}
\keyword{datasets}
