% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilde_graph.R
\name{tilde_graph}
\alias{tilde_graph}
\title{One-step data graphics}
\usage{
tilde_graph(
  D,
  tilde,
  ...,
  data = NULL,
  seed = 101,
  annot = c("none", "violin", "model"),
  jitter = c("all", "x", "y", "none"),
  model_alpha = 0.2,
  palette = LETTERS[1:8]
)
}
\arguments{
\item{D}{a data frame}

\item{tilde}{tilde expression specifying \code{y ~ x} or \code{y ~ x + color}}

\item{...}{Graphical options for the data points, e.g. alpha, size}

\item{seed}{(optional) random seed for jittering}

\item{annot}{Statistical annotation (one of "none", "violin", "model")}

\item{jitter}{Options for turning off jitter: one of \code{"none"}, \code{"x"}, \code{"y"}. By default,
categorical variables are jittered.}

\item{model_alpha}{Transparency for the model annotation}

\item{palette}{Depending on taste and visual capabilities, some people might
prefer to alter the color scheme. There are 8 palettes available: \code{"A"} through \code{"H"}.}
}
\description{
\code{tilde_graph()} makes it easy to construct an informative basic graph of a
data frame. "Making it easy" means that the user only needs to specify
two things: 1) the data frame to be used and 2) a tilde expression with the response variable on the left and up to
three explanatory variables on the right. The response variable is mapped to
the vertical axis while the first explanatory variable defines
the horizontal axis. The second explanatory variable (if any) maps to color,
the third (if any) defines facets. Quantitative variables used for color or faceting
are cut into categorical variables, so color and facets will always be discrete.
}
\details{
When an x- or y- variables is categorical, jittering is automatically applied.

Using \code{annot = "model"} will annotate the data with the graph of a
model --- shown as confidence intervals/bands --- corresponding to
the tilde expression. \code{annot = "violin"} will annotate with a violin plot.

If you want to use the same explanatory variable for color and faceting
(this might have pedagogical purposes) merely repeat the name of the color variable
in the faceting position, e.g. \code{mpg ~ hp + cyl + cyl}.
}
\examples{
Galton |> tilde_graph(height ~ mother + sex + father, annot="model", model_alpha=1)
mtcars |> tilde_graph(mpg ~ wt + cyl)
mtcars |> tilde_graph(mpg ~ wt + cyl + hp, annot="model")
}
