% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_wildfires.R
\docType{data}
\name{US_wildfires}
\alias{US_wildfires}
\title{Monthly tallies of wildfires in the US from 2000 to 2022}
\format{
275 rows, each of which is a month
\itemize{
\item \code{date} The year and month in a format that can be easily plotted
\item \code{area}: total area burned by the wildfires in that month (acres)
\item \code{number}: the number of wildfires in that month
\item \code{month}: for convenience, the month (Jan, Feb, ..., Dec) as an ordered factor.
}
}
\source{
USGS
}
\usage{
data(AAUP)
}
\description{
Records for each month of wildfires in the US.
}
\keyword{datasets}
