% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{explanatory_vars}
\alias{explanatory_vars}
\alias{response_var}
\alias{response_values}
\alias{formula_from_mod}
\title{Utilities}
\usage{
explanatory_vars(model, ...)

response_var(model, ...)

response_values(model, ...)

formula_from_mod(model, ...)
}
\arguments{
\item{model}{the model in question}

\item{...}{(not used)}
}
\description{
These are mostly pulled from the original mosaicModel package
}
