% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vertebrates.R
\docType{data}
\name{Vertebrates}
\alias{Vertebrates}
\title{Adult body mass for many species}
\format{
A data.frame with 3971 rows, each for a different species.
\itemize{
\item bodymass Mass of the adult in grams.
\item phylogentic identifiers
\item class
\item clade
\item order
\item family
\item binomial
}
}
\source{
Meiri, S., Murali, G., Zimin, A. \emph{et al.} (2021) "Different solutions lead to similar life history traits across the great divides of the amniote tree of life" \emph{J of Biol Res-Thessaloniki} \strong{28}(3) https://doi.org/10.1186/s40709-021-00134-9
}
\usage{
data("Vertebrates")
}
\description{
Body mass of adult of numerous vertebrate species. Attention was drawn to these data by a
\href{https://lsinks.github.io/posts/2023-04-03-chart-challenge-3/day3.html}{blog post} by
Louise E. Sinks
}
\keyword{datasets}
