% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_cities.R
\docType{data}
\name{World_cities}
\alias{World_cities}
\title{Location and population of large cities in the world.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 23018 rows and 10 columns.
}
\usage{
data(World_cities)
}
\description{
These data are from \code{maps::world.cities}, but included in \code{{LST}} to
avoid having to take a dependency on the \code{{maps}} package.
}
\references{
See the \code{{maps}} R package
}
\keyword{datasets}
