% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_one.R
\name{zero_one}
\alias{zero_one}
\title{Zero-one transformation for categorical variable}
\usage{
zero_one(x, one)
}
\arguments{
\item{x}{a categorical variable}

\item{one}{character string specifying the level that gets mapped to 1.}
}
\description{
A convenience function for handling categorical response variables.
Ordinarily, ggplot2 maps categorical levels to numerical values 1, 2, ....
Such numerical mapping is inappropriate for logistic modeling, where we want the levels to
be on a probability scale.
}
\examples{
Whickham |> point_plot(zero_one(outcome, one="Alive") ~ age + smoker, annot = "model")

}
