% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_worksheets.R
\name{zip_worksheets}
\alias{zip_worksheets}
\title{Zip the worksheet files}
\usage{
zip_worksheets()
}
\description{
This utility enables multiple worksheets on Posit.cloud to be downloaded
all at once in zip format
}
