% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_draw.R
\name{dag_draw}
\alias{dag_draw}
\title{Draw a DAG}
\usage{
dag_draw(DAG, ..., report_hidden = FALSE)
}
\arguments{
\item{DAG}{The DAG to draw}

\item{\ldots}{Additional arguments to plot.igraph()}

\item{show_hidden}{logical. If \code{TRUE}, show the hidden nodes.}
}
\description{
Make a simple drawing of a DAG.
}
\details{
See the igraph package for more details.

By default, edges are not drawn to hidden nodes, that is, those whose
names begin with a dot. To show the hidden
nodes, use the argument \code{show_hidden=TRUE}.
}
\examples{
dag_draw(sim_03)
}
