% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{datasim_make}
\alias{datasim_make}
\alias{datasim_run}
\alias{categorical}
\alias{cat2value}
\alias{bernoulli}
\alias{block_by}
\alias{datasim_to_igraph}
\title{Facility to make and run data simulations}
\usage{
datasim_make(...)

datasim_run(sim, n = 5, seed = NULL)

categorical(n = 5, ..., exact = TRUE)

cat2value(variable, ...)

bernoulli(n = 0, logodds = NULL, prob = 0.5, labels = NULL)

block_by(block_var, levels = c("treatment", "control"), show_block = FALSE)

datasim_to_igraph(sim, show_hidden = FALSE)
}
\arguments{
\item{sim}{A data simulation as made by \code{datasim_make()}}

\item{n}{The sample size}

\item{exact}{Logical flag. If \code{TRUE}, be scrupulous in dividing up the number of factors}

\item{variable}{a categorical variable}

\item{logodds}{Numerical vector used to generate bernouilli trials. Can be any real number.}

\item{prob}{An alternative to \code{logodds}. Values must be in \link{0,1}.}

\item{labels}{Character vector: names for categorical levels, also used to
replace 0 and 1 in bernouilli()}

\item{block_var}{Which variable to use for blocking}

\item{levels}{Character vector giving names to the blocking levels}

\item{show_block}{Logical. If \code{TRUE}, put the block number in the output.}

\item{show_hidden}{In graphing a dag, show the hidden nodes (nodes whose name begins with \code{.})}

\item{\ldots}{assignments of values to the names in \code{variable}}
}
\description{
Facility to make and run data simulations
}
