% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_one.R
\name{label_zero_one}
\alias{label_zero_one}
\title{not clear how to use this. I want an easy way to label the vertical axis of graphs
where the y-axis is a zero-one variable.}
\usage{
label_zero_one(P)
}
\arguments{
\item{P}{A ggplot2 object made by \code{model_plot()} or \code{point_plot()}}
}
\description{
not clear how to use this. I want an easy way to label the vertical axis of graphs
where the y-axis is a zero-one variable.
}
\examples{
Whickham |>
  point_plot(zero_one(outcome, one = "Alive") ~ age + smoker, annot = "model") |>
  label_zero_one() +
  ylab("Prob. alive at 20-year follow-up")
}
