% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McClave_Sincich.R
\docType{data}
\name{McClave_Sincich}
\alias{McClave_Sincich}
\alias{Clock_auction}
\alias{Geography_journals}
\alias{PGA_index}
\alias{Dowsing}
\title{Data from McClave-Sincich \emph{Statistics} 11/e}
\format{
\code{Clock_auction}: Prices for grandfather clocks sold in auction
\itemize{
\item Sales price for the clock
\item Age of the clock
\item Number of bidders for the clock
}

\code{Geography_journals}: Prices for geography journals, c. 2005
\itemize{
\item \code{journal} name of journal
\item \code{cost} for a one-year subscription
\item \code{jif} journal impact factor
\item \code{cites} number of citations of the journal in the past five years
\item \code{rpi} relative price index
}

\code{PGA_index}: Driving distance, accuracy, and a derived index from the PGA tour
\itemize{
\item \code{player} name of the player
\item \code{dist} driving distance in yards
\item \code{accuracy} percent of drives that land in the fairway
\item \code{index} an index score for ranking players.
}

\code{Dowsing}: Locations identified by dowsers in an experiment
\itemize{
\item \code{trial} just the row number
\item \code{subject} identifying number assigned to the subject
\item \code{pipe} location of the flowing-water pipe along a 10-meter line (decimeters)
\item \code{guess} the dowser's guess of the location of the pipe in that trial (decimeters)
}
}
\source{
\href{https://www.statcrunch.com/books/?book=mcclave_mstat12e}{StatCrunch}
}
\usage{
Clock_auction

Geography_journals

PGA_index

Dowsing
}
\description{
These are relatively small data frames used for exercises
}
\keyword{datasets}
