% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plot.R
\name{model_plot}
\alias{model_plot}
\title{Plot model output for representative levels of explanatory variables}
\usage{
model_plot(
  mod,
  x,
  color = NULL,
  facet = NULL,
  facet2 = NULL,
  ...,
  data = NULL,
  nlevels = 3,
  show_data = TRUE,
  interval = c("none", "prediction", "confidence"),
  data_alpha = 0.25
)
}
\description{
Plot model output for representative levels of explanatory variables
}
\examples{
mod <- lm(time ~ distance*sex*climb, data=Hill_racing)
model_plot(mod, x=climb, color=sex, facet=distance, nlevels=4, show_data=TRUE)
}
