% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_skeleton.R
\name{model_skeleton}
\alias{model_skeleton}
\title{Convert a model or a data frame to a skeleton}
\usage{
model_skeleton(mod, data = NULL, ncont = 3)
}
\arguments{
\item{mod}{A fitted model \strong{or} a tilde expression describing a model structure, e.g. \code{outcome ~ vara+varb}.}

\item{data}{a data frame. Relevant only when \code{mod} is a tilde expression}

\item{ncont}{minimum number of levels at which to represent continuous variables. (More levels
may be added to "prettify" the choices. See \code{\link[=pretty]{pretty()}}.)}
}
\description{
Only explanatory variables are included in the result.
}
