% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Samples from various kinds of objects}
\usage{
sample(x, n = 5, size = n, replace = FALSE, ...)
}
\arguments{
\item{x}{The object from which to sample}

\item{n}{An alias for \code{size}}

\item{size}{The sample size.}

\item{replace}{Logical flag: whether to sample with replacement. (default: \code{FALSE})}

\item{.by}{Variables to use to define groups for sampling, as in \code{{dplyr}}. The sample size
applies to each group.}
}
\description{
A set of methods to generate random samples from data frames and data simulations.
}
\details{
These are based in spirit on the sample functions in the \code{{mosaic}} package,
but are redefined here to 1) avoid a dependency on \code{{mosaic}} and 2) bring the arguments in
line with the \verb{.by =} features of \code{{dplyr}}.
}
