% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{explanatory_vars}
\alias{explanatory_vars}
\alias{response_var}
\alias{response_values}
\alias{formula_from_mod}
\alias{get_training_data}
\title{Utilities}
\usage{
explanatory_vars(model, ...)

response_var(model, ...)

response_values(model, ...)

formula_from_mod(model, ...)

get_training_data(model, ...)
}
\arguments{
\item{model}{the model in question}

\item{...}{(not used)}
}
\description{
Functions for pulling various components from model objects.
These work mainly for lm and glm objects. It's a future project
to add facilities for other object types.
}
