% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wheat.R
\docType{data}
\name{Wheat}
\alias{Wheat}
\title{Experimental data on the yield of winter wheat}
\format{
A data.frame with 240 rows
\itemize{
\item bodymass Mass of the adult in grams.
\item phylogentic identifiers
\item class
\item clade
\item order
\item family
\item binomial
}
}
\source{
Andrea Onofri \href{https://www.statforbiology.com/2023/stat_lmm_perennialcrops/}{"Repeated measures with perennial crops"}
}
\usage{
data("Wheat")
}
\description{
In the experiment, eight different varieties of winter wheat were planted in each of 7 calendar years (1996-2002).
}
\keyword{datasets}
